Refactoring

+++ internal/domain/repositiories.go
@@ -16,7 +16,7 @@ type Repository interface {
 
 type RepositoryUser interface {
 	Create(user User) (newUser User, err error)
-	Delete(id uint) (success bool, err error)
+	Delete(id uint64) (success bool, err error)
 	GetByEmail(email string) (user User, err error)
 	GetByUsername(username string) (user User, err error)
 	Verify(user User) error
@@ -24,12 +24,12 @@ type RepositoryUser interface {
 }
 
 type RepositoryVerifyAccount interface {
-	Create(token string, userId uint) (verification VerifyAccount, err error)
+	Create(token string, userId uint64) (verification VerifyAccount, err error)
 	Verify(token string) (verification VerifyAccount, err error)
 }
 
 type RepositoryForgetPassword interface {
-	Create(token string, userId uint) (request ForgetPassword, err error)
+	Create(token string, userId uint64) (request ForgetPassword, err error)
 	Delete(token string) (request ForgetPassword, err error)
 }
